Refactor pool/worker connection
