+++ testplan/common/entity/base.py
@@ -383,6 +383,8 @@ class Entity(logger.Loggable):
         """
         Default abort policy. First abort all dependencies and then itself.
         """
+        if not self.active:
+            return
         self._should_abort = True
         for dep in self.abort_dependencies():
             self._abort_entity(dep)
